+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## V 1.3.0 : XXX Update - XX/XX/19
 
-    - Window : Add is_running function
+    - Window : Add is_running property
     - Window-World : Rename and move WorldCallbacks to WindowCallbacks
     - WindowCallbacks : Add CHANGEWORLD and STOPWINDOW
     - Entity : Can't have the same type of component two times